refactor: merge daemons and daemon modules
