+++ client/src/components/groups/group/group.page.scss
@@ -13,75 +13,81 @@ $normal: 400;
 $bold: 600;
 $br: 4px;
 
-.group-details {
+.group-page {
 
-  background: rgba($form-bg, .9);
-  margin: 2em;
-  max-width: 70%;
-  min-width: 450px;
-  padding: 2em 5em;
-  width: 70%;
-
-  h1 {
-    color: $white;
-    font-weight: $thin;
-    margin: 0 0 30px;
-    text-transform: uppercase;
-
-    .right-floated {
-      float: right !important;
-    }
+  .ui.active.dimmer {
+      background-color: transparent;
   }
 
-  a {
-    color: $main;
-    text-decoration: none;
-    transition: .5s ease;
-    &:hover {
-      color: $main-dark;
-      cursor: pointer;
-    }
-  }
+  .group-details {
+    background: rgba($form-bg, .9);
+    margin: 2em;
+    max-width: 70%;
+    min-width: 450px;
+    padding: 2em 5em;
+    width: 70%;
 
-  .button-form {
-    .ui.button {
-      -moz-appearance: none; // Firefox
-      -webkit-appearance: none; // Safari and Chrome
-      appearance: none;
-      background: $main;
-      border: none;
-      border-radius: 0;
+    h1 {
       color: $white;
-      font-size: 1.5rem;
-      font-weight: $bold;
-      letter-spacing: .1em;
-      outline: none;
-      padding: 15px 0;
+      font-weight: $thin;
+      margin: 0 0 30px;
       text-transform: uppercase;
-      transition: all.5s ease;
-      &:hover,
-      &:focus {
-        background: $main-dark;
-        color: $white;
+
+      .right-floated {
+        float: right !important;
       }
     }
-  }
 
-  .ui.label.form-label {
-    background-color: #3a6191;
-    border-radius: 0;
-    color: white;
-    height: 38px;
-    line-height: 38px;
-    margin-top: 24px;
-    padding-bottom: 0;
-    padding-top: 0;
-    text-align: center;
-    width: 100%;
-    @media only screen and (max-width: 767px) {
-      margin-left: 25%;
-      margin-top: 0;
-      width: 50% !important;
+    a {
+      color: $main;
+      text-decoration: none;
+      transition: .5s ease;
+      &:hover {
+        color: $main-dark;
+        cursor: pointer;
+      }
+    }
+
+    .button-form {
+      .ui.button {
+        -moz-appearance: none; // Firefox
+        -webkit-appearance: none; // Safari and Chrome
+        appearance: none;
+        background: $main;
+        border: none;
+        border-radius: 0;
+        color: $white;
+        font-size: 1.5rem;
+        font-weight: $bold;
+        letter-spacing: .1em;
+        outline: none;
+        padding: 15px 0;
+        text-transform: uppercase;
+        transition: all.5s ease;
+        &:hover,
+        &:focus {
+          background: $main-dark;
+          color: $white;
+        }
+      }
+    }
+
+    .ui.label.form-label {
+      background-color: #3a6191;
+      border-radius: 0;
+      color: white;
+      height: 38px;
+      line-height: 38px;
+      margin-top: 24px;
+      padding-bottom: 0;
+      padding-top: 0;
+      text-align: center;
+      width: 100%;
+      @media only screen and (max-width: 767px) {
+        margin-left: 25%;
+        margin-top: 0;
+        width: 50% !important;
+      }
     }
   }
 }
