+++ client/src/components/services/service/service.page.scss
@@ -14,12 +14,12 @@ $bold: 600;
 $br: 4px;
 
 .service-page {
+
   .ui.active.dimmer {
       background-color: transparent;
   }
 
   .service-details {
-
     background: rgba($form-bg, .9);
     margin: 2em;
     max-width: 70%;
