--- client/src/modules/daemons/daemon/daemon.constants.js
@@ -1,9 +1,0 @@
-export default {
-  REQUEST_DAEMON : 'REQUEST_DAEMON',
-  RECEIVE_DAEMON : 'RECEIVE_DAEMON',
-  REQUEST_SAVE_DAEMON : 'REQUEST_SAVE_DAEMON',
-  DAEMON_SAVED : 'DAEMON_SAVED',
-  REQUEST_DELETE_DAEMON: 'REQUEST_DELETE_DAEMON',
-  DAEMON_DELETED: 'DAEMON_DELETED',
-  INVALID_REQUEST_DAEMON : 'INVALID_REQUEST_DAEMON'
-};
