+++ client/src/modules/daemons/daemons.constants.js
@@ -5,5 +5,12 @@ export default {
   REQUEST_DAEMON_INFO: 'REQUEST_DAEMON_INFO',
   RECEIVE_DAEMON_INFO: 'RECEIVE_DAEMON_INFO',
   INVALID_REQUEST_DAEMON_INFO: 'INVALID_REQUEST_DAEMON_INFO',
-  CHANGE_FILTER: 'CHANGE_FILTER_DAEMONS'
+  CHANGE_FILTER: 'CHANGE_FILTER_DAEMONS',
+  REQUEST_DAEMON : 'REQUEST_DAEMON',
+  RECEIVE_DAEMON : 'RECEIVE_DAEMON',
+  REQUEST_SAVE_DAEMON : 'REQUEST_SAVE_DAEMON',
+  DAEMON_SAVED : 'DAEMON_SAVED',
+  REQUEST_DELETE_DAEMON: 'REQUEST_DELETE_DAEMON',
+  DAEMON_DELETED: 'DAEMON_DELETED',
+  INVALID_REQUEST_DAEMON : 'INVALID_REQUEST_DAEMON'
 };
