+++ client/src/modules/daemons/daemons.reducer.js
@@ -1,6 +1,5 @@
 // import constants
 import DaemonsConstants from './daemons.constants.js';
-import DaemonConstants from './daemon/daemon.constants.js';
 import { generateEntitiesReducer } from '../utils/entities.js';
 
 const createRequestDaemonInfo  = (state, action) => {
@@ -49,20 +48,80 @@ const daemonsReducer = (state, action) => {
     return { ...entitiesState, ...createReceiveDaemonInfo(state, action) };
   case DaemonsConstants.INVALID_REQUEST_DAEMON_INFO:
     return { ...entitiesState, ...createInvalidDaemonInfo(state, action) };
-  case DaemonConstants.DAEMON_SAVED:
-    let items = { ...entitiesState.items };
-    items[action.daemon.id] = { ...items[action.daemon.id], ...action.daemon };
-    entitiesState.items = items;
-    return entitiesState;
-  case DaemonConstants.DAEMON_DELETED:
+  case DaemonsConstants.CHANGE_FILTER:
+    return { ...entitiesState, filterValue: action.filterValue };
+  case DaemonsConstants.INVALID_REQUEST_DAEMON:
+    return {
+      ...entitiesState,
+      selected : {
+        ...entitiesState.selected,
+        isFetching: false,
+        didInvalidate: true,
+        id: ''
+      }
+    };
+  case DaemonsConstants.REQUEST_DAEMON:
+    return {
+      ...entitiesState,
+      selected : {
+        ...entitiesState.selected,
+        isFetching: true,
+        didInvalidate: false,
+        id: action.id
+      }
+    };
+  case DaemonsConstants.RECEIVE_DAEMON:
+    const newReceivedDaemon = action.daemon;
+    const oldDaemonReceived = entitiesState.items[newReceivedDaemon.id];
+    return {
+      ...entitiesState,
+      items: {
+        ...entitiesState.items,
+        [newReceivedDaemon.id]: { ...oldDaemonReceived, ...newReceivedDaemon }
+      },
+      selected: {
+        ...entitiesState.selected,
+        isFetching: false
+      }
+    };
+  case DaemonsConstants.REQUEST_SAVE_DAEMON:
+    return {
+      ...entitiesState,
+      selected: {
+        ...entitiesState.selected,
+        isFetching: true,
+        didInvalidate: false
+      }
+    };
+  case DaemonsConstants.DAEMON_SAVED:
+    const newSavedDaemon = action.daemon;
+    const oldDaemonSaved = entitiesState.items[newSavedDaemon.id];
+    let newDaemonState = {
+      ...entitiesState,
+      items: {
+        ...entitiesState.items,
+        [newSavedDaemon.id]: { ...oldDaemonSaved, ...newSavedDaemon }
+      },
+      selected : {
+        ...entitiesState.selected,
+        isFetching: false,
+        didInvalidate: false,
+        id: newSavedDaemon.id
+      }
+    };
+    return newDaemonState;
+  case DaemonsConstants.DAEMON_DELETED:
     let deletedDaemonState = {
       ...entitiesState,
-      items: { ...entitiesState.items }
+      items: { ...entitiesState.items },
+      selected : {
+        isFetching: false,
+        didInvalidate: true,
+        id : ''
+      }
     };
     delete deletedDaemonState.items[action.id];
     return deletedDaemonState;
-  case DaemonsConstants.CHANGE_FILTER:
-    return { ...entitiesState, filterValue: action.filterValue };
   default:
     return entitiesState;
   }
